+++ rust/crd/src/user_info_fetcher.rs
@@ -1,3 +1,5 @@
+use std::collections::BTreeMap;
+
 use derivative::Derivative;
 use serde::{Deserialize, Serialize};
 use stackable_operator::{
@@ -30,6 +32,10 @@ pub enum Backend {
     /// Backend that fetches user information from the Gaia-X
     /// Cross Federation Services Components (XFSC) Authentication & Authorization Service.
     ExperimentalXfscAas(AasBackend),
+
+    /// Backend that fetches user information from Active Directory
+    #[serde(rename = "experimentalActiveDirectory")]
+    ActiveDirectory(ActiveDirectoryBackend),
 }
 
 impl Default for Backend {
@@ -86,6 +92,27 @@ pub struct AasBackend {
 
 fn aas_default_port() -> u16 {
     5000
+}
+
+#[derive(Clone, Debug, Deserialize, Eq, JsonSchema, PartialEq, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ActiveDirectoryBackend {
+    /// Hostname of the domain controller, e.g. `ad-ds-1.contoso.com`.
+    pub ldap_server: String,
+
+    /// The root Distinguished Name (DN) where users and groups are located.
+    pub base_distinguished_name: String,
+
+    /// The name of the Kerberos SecretClass.
+    pub kerberos_secret_class_name: String,
+
+    /// Use a TLS connection. If not specified then no TLS will be used.
+    #[serde(flatten)]
+    pub tls: TlsClientDetails,
+
+    /// Custom attributes, and their LDAP attribute names.
+    #[serde(default)]
+    pub custom_attribute_mappings: BTreeMap<String, String>,
 }
 
 #[derive(Clone, Debug, Deserialize, Eq, JsonSchema, PartialEq, Serialize, Derivative)]